ztest: mock: fix UNNECESSARY_INT compliance issue
Fix checkpatch issue: UNNECESSARY_INT: Prefer 'unsigned long' over
'unsigned long int' as the int is unnecessary

+++ subsys/testsuite/ztest/src/ztest_mock.c
@@ -78,7 +78,7 @@ int snprintk(char *str, size_t size, const char *fmt, ...)
  */
 #define BITS_PER_UL (8 * sizeof(unsigned long int))
 #define DEFINE_BITFIELD(name, bits)                                            \
-	unsigned long int(name)[DIV_ROUND_UP(bits, BITS_PER_UL)]
+	unsigned long(name)[DIV_ROUND_UP(bits, BITS_PER_UL)]
 
 static inline int sys_bitfield_find_first_clear(const unsigned long *bitmap,
 						const unsigned int bits)
